refactor(data-model): rename residual fabric-native-instances.ts to native-instance-utils.ts
After the by-class decomposition, the residual file holds no instance
classes anymore -- only helpers about native instances
(`isConvertibleNativeInstance`, `UNSAFE_KEYS`, `errorClassFromType`).
Rename it to match its content (Dan's call), and repoint its
`deno.json` export key + its consumers:

- `deno.json`: `./fabric-native-instances` -> `./native-instance-utils`
- `src/fabric-value-modern.ts` and `src/fabric-instances/FabricError.ts`
  (the two production consumers of the helpers)
- `test/fabric-regexp.test.ts` and `test/fabric-native-instances.test.ts`
  (the two test consumers of `isConvertibleNativeInstance`)

The file content is unchanged (git R100). No external package imported
the old `./fabric-native-instances` specifier, so the key rename is safe.
The `fabric-native-instances.test.ts` filename is left as-is (Dan asked
to rename the source residual, not the test).
