+++ packages/data-model/src/native-instance-utils.ts
