+++ packages/data-model/test/fabric-regexp.test.ts
@@ -9,7 +9,7 @@ import {
 import { BaseReconstructionContext } from "../src/BaseReconstructionContext.ts";
 import { FabricNativeWrapper } from "../src/fabric-instances/FabricNativeWrapper.ts";
 import { FabricRegExp } from "../src/fabric-instances/FabricRegExp.ts";
-import { isConvertibleNativeInstance } from "../src/fabric-native-instances.ts";
+import { isConvertibleNativeInstance } from "../src/native-instance-utils.ts";
 import {
   isFabricCompatible,
   resetDataModelConfig,
